Add medical administration record printing page

+++ app/controllers/admin/LocationsController.php
@@ -8,8 +8,6 @@ use Illuminate\Support\Facades\Input;
 use Illuminate\Support\Facades\Response;
 
 class LocationsController extends \SecuredController {
-  protected $permitted = 'admin';
-
 	/**
 	 * Display a listing of the resource.
 	 *
@@ -17,7 +15,23 @@ class LocationsController extends \SecuredController {
 	 */
 	public function index()
 	{
-    return Location::all()->toJson();
+    if ($this->permit('admin') || $this->permit('printer')) {
+      $jobs = Location::with('printJobs')->get();
+      $transformed = array();
+      foreach ($jobs as $job) {
+        $raw = $job->toArray();
+        unset($raw['print_jobs']);
+        if (!is_null($job->lastPrintJob())) {
+          $augment = array_merge((array)$raw, (array)array('last_print_job' => $job->lastPrintJob()->toArray()));
+          $transformed = array_merge((array)$transformed, (array)array($augment));  
+        } else {
+          $transformed = array_merge((array)$transformed, (array)array($raw));
+        }
+      }
+      return Response::json($transformed, 200);
+    } else {
+      return $this->unauthorizedResponse();
+    }
 	}
 
 	/**
@@ -27,10 +41,14 @@ class LocationsController extends \SecuredController {
 	 */
 	public function store()
 	{
-    $attrs = Input::all();
-		$location = Location::create($attrs);
-    Log::info("The newly created location is: ", $location->toArray());
-    Response::json($location->toJson(), 201);
+    if ($this->permit('admin')) {
+      $attrs = Input::all();
+  		$location = Location::create($attrs);
+      Log::info("The newly created location is: ", $location->toArray());
+      Response::json($location->toJson(), 201);
+    } else {
+      return $this->unauthorizedResponse();
+    }
 	}
 
 	/**
@@ -41,12 +59,16 @@ class LocationsController extends \SecuredController {
 	 */
 	public function update($id)
 	{
-    $attrs = Input::all();
-		$location = Location::find($id);
-    if ($location->update($attrs)) 
-      Response::json($location->toJson(), 201);
-    else
-      return Response::json(array('message' => 'Location failed to be updated'), 400);
+    if ($this->permit('admin')) {
+      $attrs = Input::all();
+  		$location = Location::find($id);
+      if ($location->update($attrs)) 
+        Response::json($location->toJson(), 201);
+      else
+        return Response::json(array('message' => 'Location failed to be updated'), 400);
+    } else {
+      return $this->unauthorizedResponse();
+    }
 	}
 
 	/**
@@ -57,11 +79,14 @@ class LocationsController extends \SecuredController {
 	 */
 	public function destroy($id)
 	{
-    $location = Location::find($id);
-    if ($location->delete())
-      return Response::json(array('message' => 'Location was deleted'), 200);
-    else
-      return Response::json(array('message' => 'Location failed to be deleted'), 400);
+    if ($this->permit('admin')) {
+      $location = Location::find($id);
+      if ($location->delete())
+        return Response::json(array('message' => 'Location was deleted'), 200);
+      else
+        return Response::json(array('message' => 'Location failed to be deleted'), 400);
+    } else {
+      return $this->unauthorizedResponse();
+    }
 	}
-
 }
